requirements.txt
langchain
langchain-community
python-dotenv
huggingface_hub
langchain-text-splitters
streamlit
langchain-groq
faiss-cpu
pymupdf